Fixed regression

+++ tests/src/algorithms.cpp
@@ -157,7 +157,7 @@ int main()
 	};
 
 	feature("Backward substitution") = []() {
-		test_sub<join_mats<all_mats<float, 3, 3>, all_mats<float, 3, 1>, all_mats<float, 3, 1>>>(
+		test_sub<join_mats<all_mats<double, 3, 3>, all_mats<double, 3, 1>, all_mats<double, 3, 1>>>(
 			"algorithm/back_sub/3x3_3x1.txt",
 			back_sub);
